- patch sixxs-aiccu port makefile not to use gnutls (it's too big)
- sync up M0N0WALL_GENERIC kernel config with current state of GENERIC
  config in FreeBSD 8.2 (most importantly: add uart driver for serial
  console support)

- move newbuild stuff to build/scripts and put the patches in the
  appropriate subdirs of build/patches

- make go.sh smarter: if it finds that it's been run from within a
  working copy, it will use the current working copy. Otherwise it
  will attempt to install subversion and check out the current repo
  automatically

- remove old kernel patches

- atareinit is not needed anymore

- update ipfstat patch and modify build scripts so that the patch is
  used in stage 2 (and not in stage 3 anymore)

- introduce environment variable "$MW_BUILDPATH" to avoid hard-coding
  /usr/m0n0wall/build82 in all the scripts

git-svn-id: http://svn.m0n0.ch/wall@440 e36fee2c-cc09-0410-a7cc-ebac5c6737de

+++ branches/freebsd8/build/scripts/2makebinaries.sh
@@ -2,9 +2,14 @@
 
 set -e
 
+if [ ! -d $MW_BUILDPATH ]; then
+	echo "\$MW_BUILDPATH is not set"
+	exit 1
+fi
+
 # set port options for ports that need user input
-		cd /usr/m0n0wall/build82/tmp
-		tar -zxvf /usr/m0n0wall/build82/freebsd8/build/newbuild/files/portoptions.tgz -C /
+		cd $MW_BUILDPATH/tmp
+		tar -zxvf $MW_BUILDPATH/freebsd8/build/files/portoptions.tgz -C /
  
 # autconf
         rm /usr/local/bin/autoconf
@@ -22,110 +27,111 @@ set -e
         ./buildconf --force
         ./configure --without-mysql --with-pear --with-openssl --enable-discard-path --enable-radius --enable-sockets --enable-bcmath
         make
-        install -s sapi/cgi/php ../../m0n0fs/usr/local/bin/
+        install -s sapi/cgi/php $MW_BUILDPATH/m0n0fs/usr/local/bin/
         cd ..
 # mini httpd
         fetch http://acme.com/software/mini_httpd/mini_httpd-1.19.tar.gz
         tar -zxf mini_httpd-1.19.tar.gz
         rm mini_httpd-1.19.tar.gz
         cd mini_httpd-1.19/
-        patch < ../../freebsd8/build/patches/packages/mini_httpd.patch
+        patch < $MW_BUILDPATH/freebsd8/build/patches/packages/mini_httpd.patch
         make
-        install -s mini_httpd ../../m0n0fs/usr/local/sbin
+        install -s mini_httpd $MW_BUILDPATH/m0n0fs/usr/local/sbin
         cd ..
 # wol
-        cd /usr/m0n0wall/build82/tmp
+        cd $MW_BUILDPATH/tmp
         fetch "http://downloads.sourceforge.net/project/wake-on-lan/wol/0.7.1/wol-0.7.1.tar.gz"
         tar -zxf wol-0.7.1.tar.gz
         cd wol-0.7.1
         ./configure --disable-nls
         make
-        install -s src/wol /usr/m0n0wall/build82/m0n0fs/usr/local/bin/
+        install -s src/wol $MW_BUILDPATH/m0n0fs/usr/local/bin/
 # ezipupdate
-        cd /usr/m0n0wall/build82/tmp
+        cd $MW_BUILDPATH/tmp
         fetch http://dyn.pl/client/UNIX/ez-ipupdate/ez-ipupdate-3.0.11b8.tar.gz
         tar -zxf ez-ipupdate-3.0.11b8.tar.gz
         cd ez-ipupdate-3.0.11b8
-        patch < ../../freebsd8/build/patches/packages/ez-ipupdate.c.patch
+        patch < $MW_BUILDPATH/freebsd8/build/patches/packages/ez-ipupdate.c.patch
         ./configure
         make
-        install -s ez-ipupdate /usr/m0n0wall/build82/m0n0fs/usr/local/bin/
+        install -s ez-ipupdate $MW_BUILDPATH/m0n0fs/usr/local/bin/
         cd ..
 # ipfilter userland tools (newer version than included with FreeBSD)
-		cd /usr/m0n0wall/build82/tmp
-		tar -zxvf /usr/m0n0wall/build82/freebsd8/build/local-sources/ip_fil4.1.34.tar.gz
+		cd $MW_BUILDPATH/tmp
+		tar -zxvf $MW_BUILDPATH/freebsd8/build/local-sources/ip_fil4.1.34.tar.gz
 		cd ip_fil4.1.34
+        patch < $MW_BUILDPATH/freebsd8/build/patches/user/ipfstat.c.patch
 		make freebsd8
-		install -s BSD/FreeBSD-8.?-RELEASE-i386/{ipf,ipfs,ipfstat,ipmon,ipnat} /usr/m0n0wall/build82/m0n0fs/sbin
+		install -s BSD/FreeBSD-8.?-RELEASE-i386/{ipf,ipfs,ipfstat,ipmon,ipnat} $MW_BUILDPATH/m0n0fs/sbin
 # ISC dhcp-relay
         cd /usr/ports/net/isc-dhcp31-relay
         make
-        install -s work/dhcp-*/work.freebsd/relay/dhcrelay /usr/m0n0wall/build82/m0n0fs/usr/local/sbin/
+        install -s work/dhcp-*/work.freebsd/relay/dhcrelay $MW_BUILDPATH/m0n0fs/usr/local/sbin/
 # ISC dhcp-server
         cd /usr/ports/net/isc-dhcp31-server
-		cp /usr/m0n0wall/build82/freebsd8/build/patches/packages/isc-dhcpd/patch-server.db.c /usr/ports/net/isc-dhcp31-server/files/
+		cp $MW_BUILDPATH/freebsd8/build/patches/packages/isc-dhcpd/patch-server.db.c /usr/ports/net/isc-dhcp31-server/files/
         make
-        install -s work/dhcp-*/work.freebsd/server/dhcpd /usr/m0n0wall/build82/m0n0fs/usr/local/sbin/
+        install -s work/dhcp-*/work.freebsd/server/dhcpd $MW_BUILDPATH/m0n0fs/usr/local/sbin/
 # ISC dhcp-client
 		cd /usr/ports/net/isc-dhcp31-client
 		make
-		install -s work/dhcp-*/work.freebsd/client/dhclient /usr/m0n0wall/build82/m0n0fs/sbin/
+		install -s work/dhcp-*/work.freebsd/client/dhclient $MW_BUILDPATH/m0n0fs/sbin/
 # dnsmasq
         cd /usr/ports/dns/dnsmasq
 		make
-        install -s work/dnsmasq-*/src/dnsmasq /usr/m0n0wall/build82/m0n0fs/usr/local/sbin/
+        install -s work/dnsmasq-*/src/dnsmasq $MW_BUILDPATH/m0n0fs/usr/local/sbin/
 # ipsec-tools
         cd /usr/ports/security/ipsec-tools
-		patch < /usr/m0n0wall/build82/freebsd8/build/newbuild/patches/ipsec-tools.Makefile.patch
+		patch < $MW_BUILDPATH/freebsd8/build/patches/packages/ipsec-tools.Makefile.patch
         make
-        install -s work/ipsec-tools-*/src/racoon/.libs/racoon /usr/m0n0wall/build82/m0n0fs/usr/local/sbin
-        install -s work/ipsec-tools-*/src/libipsec/.libs/libipsec.so.0 /usr/m0n0wall/build82/m0n0fs/usr/local/lib
+        install -s work/ipsec-tools-*/src/racoon/.libs/racoon $MW_BUILDPATH/m0n0fs/usr/local/sbin
+        install -s work/ipsec-tools-*/src/libipsec/.libs/libipsec.so.0 $MW_BUILDPATH/m0n0fs/usr/local/lib
 # dhcp6
 		cd /usr/ports/net/dhcp6
 		make
-		install -s work/wide-dhc*/dhcp6c /usr/m0n0wall/build82/m0n0fs/usr/local/sbin
-		install -s work/wide-dhc*/dhcp6s /usr/m0n0wall/build82/m0n0fs/usr/local/sbin
+		install -s work/wide-dhc*/dhcp6c $MW_BUILDPATH/m0n0fs/usr/local/sbin
+		install -s work/wide-dhc*/dhcp6s $MW_BUILDPATH/m0n0fs/usr/local/sbin
 # sixxs-aiccu		
 		cd /usr/ports/net/sixxs-aiccu/
+		patch < $MW_BUILDPATH/freebsd8/build/patches/packages/sixxs-aiccu.Makefile.patch
 		make
-		install -s work/aiccu/unix-console/aiccu /usr/m0n0wall/build82/m0n0fs/usr/local/sbin/sixxs-aiccu
+		install -s work/aiccu/unix-console/aiccu $MW_BUILDPATH/m0n0fs/usr/local/sbin/sixxs-aiccu
 # rtadvd		
 		cd /usr/src/usr.sbin/rtadvd
 		make
-		install -s rtadvd /usr/m0n0wall/build82/m0n0fs/usr/sbin/
+		install -s rtadvd $MW_BUILDPATH/m0n0fs/usr/sbin/
 # mpd4
 		cd /usr/ports/net/mpd4
 		make
-		install -s work/mpd-4.4.1/src/mpd4 /usr/m0n0wall/build82/m0n0fs/usr/local/sbin/
+		install -s work/mpd-4.4.1/src/mpd4 $MW_BUILDPATH/m0n0fs/usr/local/sbin/
 # mbmon
 		cd /usr/ports/sysutils/mbmon
 		make
-		install -s work/xmbmon205/mbmon /usr/m0n0wall/build82/m0n0fs/usr/local/bin/
+		install -s work/xmbmon205/mbmon $MW_BUILDPATH/m0n0fs/usr/local/bin/
 
 # make m0n0wall tools and binaries
-        cd /usr/m0n0wall/build82/tmp
-        cp -r ../freebsd8/build/tools .
+        cd $MW_BUILDPATH/tmp
+        cp -r $MW_BUILDPATH/freebsd8/build/tools .
         cd tools
         gcc -o stats.cgi stats.c
         gcc -o minicron minicron.c
-        # gcc -o atareinit atareinit.c
         gcc -o choparp choparp.c
         gcc -o verifysig -lcrypto verifysig.c
         gcc -o dnswatch dnswatch.c
-        install -s choparp /usr/m0n0wall/build82/m0n0fs/usr/local/sbin
-        install -s stats.cgi /usr/m0n0wall/build82/m0n0fs/usr/local/www
-        install -s minicron /usr/m0n0wall/build82/m0n0fs//usr/local/bin
-        install -s verifysig /usr/m0n0wall/build82/m0n0fs/usr/local/bin
-        install -s dnswatch /usr/m0n0wall/build82/m0n0fs/usr/local/bin
-        install runsntp.sh /usr/m0n0wall/build82/m0n0fs/usr/local/bin
-        install ppp-linkup vpn-linkdown vpn-linkup /usr/m0n0wall/build82/m0n0fs/usr/local/sbin
+        install -s choparp $MW_BUILDPATH/m0n0fs/usr/local/sbin
+        install -s stats.cgi $MW_BUILDPATH/m0n0fs/usr/local/www
+        install -s minicron $MW_BUILDPATH/m0n0fs//usr/local/bin
+        install -s verifysig $MW_BUILDPATH/m0n0fs/usr/local/bin
+        install -s dnswatch $MW_BUILDPATH/m0n0fs/usr/local/bin
+        install runsntp.sh $MW_BUILDPATH/m0n0fs/usr/local/bin
+        install ppp-linkup vpn-linkdown vpn-linkup $MW_BUILDPATH/m0n0fs/usr/local/sbin
 
 # ucd-snmp
         rm /usr/local/bin/autoconf
         rm /usr/local/bin/autoheader
         ln -s /usr/local/bin/autoconf-2.62 /usr/local/bin/autoconf
         ln -s /usr/local/bin/autoheader-2.62 /usr/local/bin/autoheader
-        cd /usr/m0n0wall/build82/tmp
+        cd $MW_BUILDPATH/tmp
         fetch "http://downloads.sourceforge.net/project/net-snmp/ucd-snmp/4.2.7.1/ucd-snmp-4.2.7.1.tar.gz"
         tar -zxf ucd-snmp-4.2.7.1.tar.gz
         cd ucd-snmp-4.2.7.1
@@ -135,6 +141,6 @@ set -e
         '--with-out-transports=TCP Unix' \
         '--with-mib-modules=mibII/interfaces mibII/var_route ucd-snmp/vmstat_freebsd2' \
 		--with-defaults
-		patch < /usr/m0n0wall/build82/freebsd8/build/newbuild/patches/ucd-snmp.config.h.patch
+		patch < $MW_BUILDPATH/freebsd8/build/patches/packages/ucd-snmp.config.h.patch
 	    make
-        install -s agent/snmpd /usr/m0n0wall/build82/m0n0fs/usr/local/sbin
+        install -s agent/snmpd $MW_BUILDPATH/m0n0fs/usr/local/sbin
